+++ app/Widgets/Vcard4/Vcard4.php
@@ -42,7 +42,6 @@ class Vcard4 extends Base
         $this->toast($this->__('vcard.updated'));
 
         $this->rpc('MovimTpl.fill', '#vcard_form', $html);
-        $this->rpc('MovimUtils.applyAutoheight');
     }
 
     public function onMyVcard4Received()